BitcoinCash: extracted work_reuired to separate file
